# openapi.yaml
openapi: 3.1.0
info:
  title: ZVERY Business Navigator API
  version: 0.1.0
  description: >
    Один контракт /api/v1. services/api слушает :8000 и отдаёт каталог.
    services/bot слушает :8001 и отдаёт пользовательские маршруты.
    Прямой :8000 не реализует bot-маршруты. GET /api/v1/measures/saved на API
    совпадает с карточкой ID saved и отвечает 404 measure_not_found.
servers:
  - url: http://127.0.0.1:8000
    description: services/api, прямой порт каталога
  - url: http://127.0.0.1:8001
    description: services/bot, прямой порт пользователя и webhook
  - url: https://zverybot.ru
    description: Публичный gateway. Маршрутизация этой фазой не меняется.
paths:
...
  /api/v1/auth/max/launch-data:
    post:
      x-zvery-service: bot
      operationId: verifyLaunchData
      servers:
        - url: http://127.0.0.1:8001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaunchDataRequest'
      responses:
        '200':
          description: Verified MAX user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaunchDataResponse'
        '401':
          description: Invalid launch data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/measures/{measure_id}/save:
    post:
      x-zvery-service: bot
      operationId: saveMeasure
      servers:
        - url: http://127.0.0.1:8001
      parameters:
        - $ref: '#/components/parameters/MeasureId'
        - $ref: '#/components/parameters/MaxInitData'
      responses:
        '200':
          description: Measure saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResult'
        '401':
          description: Launch data rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Measure id is not in the catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      x-zvery-service: bot
      operationId: removeSavedMeasure
      servers:
        - url: http://127.0.0.1:8001
      parameters:
        - $ref: '#/components/parameters/MeasureId'
        - $ref: '#/components/parameters/MaxInitData'
      responses:
        '200':
          description: Measure unsaved. Repeated delete is idempotent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveSavedResult'
        '401':
          description: Launch data rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/measures/saved:
    get:
      x-zvery-service: bot
      operationId: listSavedMeasures
      servers:
        - url: http://127.0.0.1:8001
      parameters:
        - $ref: '#/components/parameters/MaxInitData'
      responses:
        '200':
          description: Saved measure ids for the verified user. Not served by API :8000.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedMeasures'
        '401':
          description: Launch data rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/quiz/submit:
    post:
      x-zvery-service: bot
      operationId: submitQuiz
      servers:
        - url: http://127.0.0.1:8001
      parameters:
        - $ref: '#/components/parameters/MaxInitData'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizSubmitRequest'
      responses:
        '200':
          description: Quiz result and optional signed certificate payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizSubmitResponse'
        '401':
          description: Launch data rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Quiz payload rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Quiz or certificate signing is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/notifications/opt-in:
    get:
      x-zvery-service: bot
      operationId: getNotificationOptIn
      servers:
        - url: http://127.0.0.1:8001
      parameters:
        - $ref: '#/components/parameters/MaxInitData'
      responses:
        '200':
          description: Global notification consent. Default is disabled. Not a reminder subscription.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptInResponse'
        '401':
          description: Launch data rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      x-zvery-service: bot
      operationId: notificationOptIn
      servers:
        - url: http://127.0.0.1:8001
      parameters:
        - $ref: '#/components/parameters/MaxInitData'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptInRequest'
      responses:
        '200':
          description: Stores the global consent flag. Does not send reminders.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptInResponse'
        '401':
          description: Launch data rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /webhooks/max:
    post:
      x-zvery-service: bot
      operationId: maxWebhook
      servers:
        - url: http://127.0.0.1:8001
      parameters:
        - $ref: '#/components/parameters/MaxWebhookSecret'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: MAX update accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAck'
        '401':
          description: Webhook secret rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Webhook secret is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    MeasureId:
      in: path
      name: measure_id
      required: true
      schema:
        type: string
    MaxInitData:
      in: header
      name: X-Max-Init-Data
      required: true
      schema:
        type: string
      description: Исходная строка запуска MAX. Пустое или неверное значение — 401.
    MaxWebhookSecret:
      in: header
      name: X-Max-Bot-Api-Secret
      required: true
      schema:
        type: string
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              description: Создаёт сервер. Заголовок X-Request-ID клиента игнорируется.
          additionalProperties: false
      additionalProperties: false
...
    LaunchDataRequest:
      type: object
      required: [init_data]
      properties:
        init_data:
          type: string
    LaunchDataResponse:
      type: object
      required: [user, auth_date]
      properties:
        user:
          type: object
          required: [id]
          properties:
            id:
              type: string
            first_name:
              type: ['string', 'null']
            last_name:
              type: ['string', 'null']
            username:
              type: ['string', 'null']
        auth_date:
          type: integer
    SaveResult:
      type: object
      required: [measure_id, saved, created]
      properties:
        measure_id:
          type: string
        saved:
          type: boolean
          const: true
        created:
          type: boolean
    RemoveSavedResult:
      type: object
      required: [measure_id, saved, removed]
      properties:
        measure_id:
          type: string
        saved:
          type: boolean
          const: false
        removed:
          type: boolean
    SavedMeasures:
      type: object
      required: [measure_ids]
      properties:
        measure_ids:
          type: array
          items:
            type: string
    QuizSubmitRequest:
      type: object
      required: [quiz_version, answers]
      properties:
        quiz_version:
          type: string
        answers:
          type: object
          additionalProperties:
            type: string
    QuizSubmitResponse:
      type: object
      required: [attempt_id, score, passed, pass_score, certificate]
      properties:
        attempt_id:
          type: string
        score:
          type: integer
        passed:
          type: boolean
        pass_score:
          type: integer
          description: Server threshold. The client must not invent one.
        certificate:
          description: Null when not passed. Never a public verification document.
          oneOf:
            - type: 'null'
            - type: object
              required: [certificate_id, title, disclaimer, payload]
              properties:
                certificate_id:
                  type: string
                title:
                  type: string
                disclaimer:
                  type: string
                payload:
                  type: string
                  description: base64url JSON plus HMAC-SHA256 hex. Not a cryptographic signature of a state document and not a public verification URL.
    OptInRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    OptInResponse:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    WebhookAck:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          const: true
        duplicate:
          type: boolean
...